I discovered this one while copying the code for `merge` (https://github.com/haskell-unordered-containers/unordered-containers/issues/226).

Not sure whether it's actually correct or useful, but tests have passed so far.

`t1` and `t2` are always either `Leaf` or `Collision`, in case that matters.

---
TODO:
* [x] Check effects on Core
* [x] Benchmark