@sjakobi 2021-04-29
> You going to unWIP this? Or still considering?

Yeah, once I've checked the Core and run the benchmarks. Feel free to take over if you're interested though.