@sjakobi 2021-11-29
I've had a look at the Core generated with GHC 8.10.7. The most significant effect is that in the `Strict` version, the `goDifferentHash` worker is floated out. Previously that was probably prevented because it was mutually recursive with `go`.

Also its `BitmapIndexed` return value is unboxed now, which seems beneficial.

The `Lazy` version doesn't have any notable changes AFAICT.

See for yourself:

This branch:
* [Internal.dump-simpl.txt](https://github.com/haskell-unordered-containers/unordered-containers/files/7618422/Internal.dump-simpl.txt)
* [Strict.dump-simpl.txt](https://github.com/haskell-unordered-containers/unordered-containers/files/7618427/Strict.dump-simpl.txt)

`master`:
* [Internal.dump-simpl.txt](https://github.com/haskell-unordered-containers/unordered-containers/files/7618431/Internal.dump-simpl.txt)
* [Strict.dump-simpl.txt](https://github.com/haskell-unordered-containers/unordered-containers/files/7618433/Strict.dump-simpl.txt)