@sjakobi 2021-11-29
In the `HashMap/union` benchmark I'm seeing a speedup of around 1% which seems significant:

`master` (2 runs):

```
benchmarked HashMap/union
time                 69.16 μs   (69.00 μs .. 69.35 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 69.29 μs   (69.23 μs .. 69.35 μs)
std dev              198.0 ns   (155.3 ns .. 259.6 ns)

benchmarked HashMap/union
time                 69.05 μs   (68.67 μs .. 69.42 μs)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 68.96 μs   (68.81 μs .. 69.24 μs)
std dev              598.9 ns   (293.3 ns .. 946.4 ns)
```

This branch (also 2 runs):

```
benchmarked HashMap/union
time                 68.24 μs   (68.10 μs .. 68.42 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 68.26 μs   (68.22 μs .. 68.32 μs)
std dev              154.3 ns   (107.2 ns .. 204.9 ns)

benchmarked HashMap/union
time                 68.10 μs   (68.05 μs .. 68.15 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 68.46 μs   (68.39 μs .. 68.54 μs)
std dev              249.1 ns   (198.1 ns .. 304.8 ns)
```

Unfortunately we don't have any benchmarks for the strict `union` variants, where I'd expect a larger impact.

Nevertheless it seems to me that all signs point in the right direction, so I'm inclined to merge.